+++ CHANGELOG.md
@@ -54,6 +54,7 @@
    * UPDATED: bump tz from 2025a to 2025b [#5164](https://github.com/valhalla/valhalla/pull/5164)
    * ADDED: Mutithreaded `PBFGraphParser::ParseWays()` [#5143](https://github.com/valhalla/valhalla/pull/5143)
    * CHANGED: "Multilevel Way" message logging level changed from WARN to DEBUG [#5188](https://github.com/valhalla/valhalla/pull/5188)
+   * Use rapidjson for matrix serializers [#5189](https://github.com/valhalla/valhalla/pull/5189)
    * CHANGED: Clip admin/languages/timezones polygons by tile bbox for faster calculus [#5193](https://github.com/valhalla/valhalla/pull/5193)
 
 ## Release Date: 2024-10-10 Valhalla 3.5.1
@@ -379,7 +380,8 @@
    * FIXED: Osmway struct update: added up to 33 and not 32 [#3808](https://github.com/valhalla/valhalla/pull/3808)
    * FIXED: Fix out-of-range linestrings in expansion [#4603](https://github.com/valhalla/valhalla/pull/4603)
    * FIXED: Osmway struct update: used 1 bit for multiple levels from spare bits [#5112](https://github.com/valhalla/valhalla/issues/5112)
-   * FIXES: multi-edge steps maneuvers [#5191](https://github.com/valhalla/valhalla/pull/5191)
+   * FIXED: multi-edge steps maneuvers [#5191](https://github.com/valhalla/valhalla/pull/5191)
+   * FIXED: remove start maneuver if route starts on stairs/escalators [#5127](https://github.com/valhalla/valhalla/pull/5127)
 
 * **Enhancement**
    * CHANGED: Pronunciation for names and destinations [#3132](https://github.com/valhalla/valhalla/pull/3132)